requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
gunicorn==22.0.0
h11==0.14.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2024.1
uvicorn==0.30.3
